Added store not found when deleting

+++ src/main/java/com/example/practice/stores/StoreService.java
@@ -42,6 +42,7 @@ public class StoreService {
     }
 
     public void deleteStoreById(Long id) {
+        if (!repository.existsById(id)) throw new StoreNotFound();
         repository.deleteById(id);
     }
 }
